python_service/tests/test_pdf_analyzer.py

import pytest
import asyncio
from pathlib import Path
import sys
sys.path.append(str(Path(__file__).parent.parent))

from app.services.pdf_analyzer import PDFAnalyzer

class TestPDFAnalyzer:
    @pytest.fixture
    def analyzer(self):
        return PDFAnalyzer()
    
    @pytest.mark.asyncio
    async def test_analyze_basic_properties(self, analyzer):
        """Temel PDF özelliklerini test eder"""
        # Test için basit bir PDF oluştur veya mevcut bir PDF kullan
        # Bu test için mock data kullanıyoruz
        pass
    
    @pytest.mark.asyncio
    async def test_dimension_analysis(self, analyzer):
        """Boyut analizi testleri"""
        pass
    
    @pytest.mark.asyncio
    async def test_content_analysis(self, analyzer):
        """İçerik analizi testleri"""
        pass
    
    @pytest.mark.asyncio
    async def test_quality_analysis(self, analyzer):
        """Kalite analizi testleri"""
        pass

if __name__ == "__main__":
    pytest.main([__file__])
