python_service/requirements.txt

fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
PyMuPDF==1.23.0
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3
pdf2image==1.16.3
reportlab==4.0.7
svgwrite==1.4.3
python-magic==0.4.27
aiofiles==23.2.1
